added feature to define user types in django settings
